Dashboard/services/fast_pdf_processor.py
"""
Fast PDF Processor with PyMuPDF4LLM
====================================
Optimized PDF processing using PyMuPDF4LLM for 10-20x speed improvement.
Falls back to table extractors only when needed.
"""

import os
import time
import logging
import hashlib
from typing import List, Dict, Any, Optional
from pathlib import Path
from datetime import datetime

logger = logging.getLogger(__name__)

# Primary parser - PyMuPDF4LLM (fastest, best for LLMs)
try:
    import pymupdf4llm
    PYMUPDF4LLM_AVAILABLE = True
    logger.info("PyMuPDF4LLM available - using fast PDF processing")
except ImportError:
    PYMUPDF4LLM_AVAILABLE = False
    logger.warning("PyMuPDF4LLM not available. Install with: pip install pymupdf4llm")

# Fallback to basic PyMuPDF if needed
try:
    import fitz  # PyMuPDF
    PYMUPDF_AVAILABLE = True
except ImportError:
    PYMUPDF_AVAILABLE = False
    logger.warning("PyMuPDF not available. Install with: pip install PyMuPDF")

# Table extractors removed - not used in production


class FastPDFProcessor:
    """
    Fast PDF processor optimized for speed and LLM compatibility.
    Uses PyMuPDF4LLM as primary parser with intelligent fallbacks.
    """

    def __init__(self):
        """
        Initialize the fast PDF processor.
        """
        if not PYMUPDF4LLM_AVAILABLE and not PYMUPDF_AVAILABLE:
            raise ImportError("Neither PyMuPDF4LLM nor PyMuPDF available. Install with: pip install pymupdf4llm")

    def process_pdf(self, pdf_path: str, chunk_size: int = 512, chunk_overlap: int = 64) -> List[Dict[str, Any]]:
        """
        Process PDF with optimized pipeline.

        Args:
            pdf_path: Path to PDF file
            chunk_size: Target chunk size (characters)
            chunk_overlap: Overlap between chunks

        Returns:
            List of chunks with metadata
        """
        pdf_path = Path(pdf_path)
        if not pdf_path.exists():
            raise FileNotFoundError(f"PDF not found: {pdf_path}")

        start_time = time.time()
        logger.info(f"Processing PDF: {pdf_path.name}")

        try:
            # Primary method: PyMuPDF4LLM (fastest, best quality)
            if PYMUPDF4LLM_AVAILABLE:
                chunks = self._process_with_pymupdf4llm(pdf_path, chunk_size, chunk_overlap)
                method = "pymupdf4llm"
            # Fallback: Basic PyMuPDF
            elif PYMUPDF_AVAILABLE:
                chunks = self._process_with_pymupdf(pdf_path, chunk_size, chunk_overlap)
                method = "pymupdf"
            else:
                raise ValueError("No PDF processing method available")

            elapsed = time.time() - start_time
            logger.info(f"Processed {len(chunks)} chunks in {elapsed:.2f}s using {method}")

            return chunks

        except Exception as e:
            logger.error(f"PDF processing failed: {e}")
            raise

    def _process_with_pymupdf4llm(self, pdf_path: Path, chunk_size: int, chunk_overlap: int) -> List[Dict[str, Any]]:
        """
        Process PDF using PyMuPDF4LLM (fastest method).

        This is 10-20x faster than Camelot/Tabula and produces better LLM-ready output.
        """
        logger.info("Using PyMuPDF4LLM for fast processing")

        # Extract markdown with tables preserved
        md_text = pymupdf4llm.to_markdown(
            str(pdf_path),
            page_chunks=False,  # Get all pages as one text for better chunking
            write_images=False,  # Skip images for speed
            show_progress=False,  # Disable progress bar for server use
        )

        # Smart chunking that respects markdown structure
        chunks = self._smart_chunk_markdown(md_text, chunk_size, chunk_overlap)

        # Add metadata to chunks
        for i, chunk in enumerate(chunks):
            chunk.update({
                'chunk_index': i,
                'source': pdf_path.name,
                'method': 'pymupdf4llm',
                'timestamp': datetime.now().isoformat()
            })

        return chunks

    def _process_with_pymupdf(self, pdf_path: Path, chunk_size: int, chunk_overlap: int) -> List[Dict[str, Any]]:
        """
        Fallback processing with basic PyMuPDF.
        """
        logger.info("Using basic PyMuPDF processing")

        doc = fitz.open(str(pdf_path))
        all_text = []

        for page_num, page in enumerate(doc):
            text = page.get_text()
            if text.strip():
                all_text.append(f"## Page {page_num + 1}\n{text}")

        doc.close()

        # Join all pages
        full_text = "\n\n".join(all_text)

        # Basic chunking
        chunks = self._basic_chunk_text(full_text, chunk_size, chunk_overlap)

        # Add metadata
        for i, chunk in enumerate(chunks):
            chunk.update({
                'chunk_index': i,
                'source': pdf_path.name,
                'method': 'pymupdf',
                'timestamp': datetime.now().isoformat()
            })

        return chunks

    def _smart_chunk_markdown(self, text: str, chunk_size: int, chunk_overlap: int) -> List[Dict[str, Any]]:
        """
        Smart chunking that respects markdown structure (headers, tables, lists).
        """
        chunks = []

        # Split by major sections (headers)
        sections = text.split('\n## ')

        current_chunk = ""

        for section in sections:
            # Add header back if not first section
            if section != sections[0]:
                section = "## " + section

            # If section fits in chunk, add it
            if len(current_chunk) + len(section) <= chunk_size:
                current_chunk += section + "\n"
            else:
                # Save current chunk if not empty
                if current_chunk:
                    chunks.append({'content': current_chunk.strip()})

                # If section itself is too large, split it
                if len(section) > chunk_size:
                    sub_chunks = self._split_large_section(section, chunk_size, chunk_overlap)
                    chunks.extend(sub_chunks)
                    current_chunk = ""
                else:
                    current_chunk = section + "\n"

        # Don't forget last chunk
        if current_chunk:
            chunks.append({'content': current_chunk.strip()})

        return chunks

    def _split_large_section(self, text: str, chunk_size: int, chunk_overlap: int) -> List[Dict[str, Any]]:
        """
        Split large sections while preserving structure.
        """
        chunks = []

        # Try to split by paragraphs first
        paragraphs = text.split('\n\n')

        current_chunk = ""

        for para in paragraphs:
            if len(current_chunk) + len(para) <= chunk_size:
                current_chunk += para + "\n\n"
            else:
                if current_chunk:
                    chunks.append({'content': current_chunk.strip()})

                # Handle overlap
                if chunks and chunk_overlap > 0:
                    overlap_text = current_chunk[-chunk_overlap:] if len(current_chunk) > chunk_overlap else current_chunk
                    current_chunk = overlap_text + para + "\n\n"
                else:
                    current_chunk = para + "\n\n"

        if current_chunk:
            chunks.append({'content': current_chunk.strip()})

        return chunks

    def _basic_chunk_text(self, text: str, chunk_size: int, chunk_overlap: int) -> List[Dict[str, Any]]:
        """
        Basic text chunking with overlap.
        """
        chunks = []
        start = 0

        while start < len(text):
            end = start + chunk_size

            # Try to break at sentence boundary
            if end < len(text):
                # Look for sentence end
                for punct in ['. ', '.\n', '! ', '!\n', '? ', '?\n']:
                    last_punct = text.rfind(punct, start, end)
                    if last_punct != -1:
                        end = last_punct + 1
                        break

            chunk_text = text[start:end].strip()
            if chunk_text:
                chunks.append({'content': chunk_text})

            # Move start position with overlap
            start = end - chunk_overlap if chunk_overlap > 0 else end

        return chunks


# Convenience function for compatibility
def create_fast_pdf_chunks(pdf_path: str, chunk_size: int = 512, chunk_overlap: int = 64) -> List[Dict[str, Any]]:
    """
    Create chunks from PDF using fast processing.

    This is a drop-in replacement for create_enhanced_pdf_chunks.
    """
    processor = FastPDFProcessor()
    return processor.process_pdf(pdf_path, chunk_size, chunk_overlap)